pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "yolo_v1"
version = "0.0.1"
dependencies = [
    "torch",
    "torchvision",
    "opencv-python",
    "numpy",
    "albumentations",
    "tqdm",
    "pandas",
    "Pillow",
    "hydra-core",
    "omegaconf",
    "wandb",
    "matplotlib",
    "scipy",
    "opencv-python",
    "PyYAML",
    "Pillow",
    "ultralytics",
    "requests",
    "psutil",
    "sentry-sdk",
    "setproctitle",
    "GitPython",
    "docker-pycreds",
    "Jinja2",
    "nvidia-cublas-cu12",
    "nvidia-cuda-cupti-cu12",
    "nvidia-cuda-nvrtc-cu12",
    "nvidia-cuda-runtime-cu12",
    "nvidia-cudnn-cu12",
    "nvidia-cufft-cu12",
    "nvidia-cufile-cu12",
    "nvidia-curand-cu12",
    "nvidia-cusolver-cu12",
    "nvidia-cusparse-cu12",
    "nvidia-cusparselt-cu12",
    "nvidia-nccl-cu12",
    "nvidia-nvjitlink-cu12",
    "nvidia-nvshmem-cu12",
    "nvidia-nvtx-cu12",
    "ruff",
    "pytest"
]

[tool.setuptools.packages]
find = {"where" = ["."], "include" = ["yolo_v1", "yolo_v1.*"]}